mensajes de error de borrar y acceder
