+++ INTERFACES/IDeleteUserView.java
@@ -4,4 +4,6 @@ public interface IDeleteUserView {
     String userToDelete();
     int sureToDelete();
     void deleteUserMsg();
+
+    void errordeleteUserMsg();
 }
